demo1/requirements.txt
streamlit
pymongo
python-dotenv
openai
flask
twilio
pyngrok